+++ lib/visitor.dart
@@ -4,6 +4,7 @@ import 'package:analyzer/analysis_rule/rule_visitor_registry.dart';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:analyzer/dart/ast/visitor.dart';
 import 'package:analyzer/dart/element/element.dart';
+import 'package:analyzer/dart/element/nullability_suffix.dart';
 import 'package:analyzer/dart/element/type.dart';
 
 import 'main.dart';
@@ -104,10 +105,70 @@ class Visitor extends SimpleAstVisitor<void> {
         Expression expr => expr,
       };
 
-      if (expression.hasArgumentExpression) {
+      if (expression.isDotShorthand) continue;
+
+      final prefixElement = expression.extractInterfaceElement();
+      if (prefixElement == null) continue;
+
+      final expressionType = expression.staticType;
+      if (expressionType == null) continue;
+
+      final parameterType = _getParameterType(argument, expression);
+      if (parameterType == null) continue;
+
+      final prefixType = prefixElement.thisType;
+      final parameterBaseType = _getNonNullableType(parameterType);
+      final expressionBaseType = _getNonNullableType(expressionType);
+
+      if (expressionBaseType == prefixType && parameterBaseType == prefixType) {
         rule.reportAtNode(expression);
       }
     }
+  }
+
+  DartType? _getParameterType(Expression argument, Expression expression) {
+    if (argument is NamedExpression) {
+      return argument.element?.type;
+    }
+
+    if (expression.parent case final parent?) {
+      if (parent is NamedExpression) {
+        return parent.element?.type;
+      }
+    }
+
+    final argumentList = expression.thisOrAncestorOfType<ArgumentList>();
+    if (argumentList == null) return null;
+
+    final parent = argumentList.parent;
+    final argumentIndex = argumentList.arguments.indexOf(argument);
+    if (argumentIndex == -1) return null;
+
+    if (parent is MethodInvocation) {
+      final element = parent.methodName.element;
+      return _getParameterTypeFromElement(element, argumentIndex);
+    } else if (parent is InstanceCreationExpression) {
+      final element = parent.constructorName.element;
+      return _getParameterTypeFromElement(element, argumentIndex);
+    }
+
+    return null;
+  }
+
+  DartType? _getParameterTypeFromElement(Element? element, int index) {
+    if (element is! ExecutableElement) return null;
+
+    final formalParameters = element.formalParameters;
+    if (index < 0 || index >= formalParameters.length) return null;
+
+    return formalParameters[index].type;
+  }
+
+  DartType _getNonNullableType(DartType type) {
+    if (type.nullabilitySuffix == NullabilitySuffix.none) {
+      return type;
+    }
+    return context.typeSystem.promoteToNonNull(type);
   }
 }
 
@@ -151,13 +212,4 @@ extension on Expression {
       e,
     _ => null,
   };
-
-  bool get hasArgumentExpression {
-    if (isDotShorthand) return false;
-
-    final prefixElement = extractInterfaceElement();
-    if (prefixElement == null) return false;
-
-    return prefixElement.isPrefixType(staticType);
-  }
 }