+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.2.0
+
+- Fix case:
+
+```dart
+class Animal {}
+class Dog {
+    static Dog husky() => ...;
+}
+
+void foo(Animal animal){}
+
+void main() {
+    foo(Dog.husky()); // should not warning
+}
+```
+
 ## 0.1.0+1
 
 - Update `README.md`.